Домашнее задание к лекции 5

+++ MetricsAgent/DAL/Repositories/DateTimeOffsetHandler.cs
@@ -17,7 +17,7 @@ namespace MetricsAgent.DAL.Repositories
 
 	    public override DateTimeOffset Parse(object value)
 	    {
-		    return DateTimeOffset.FromUnixTimeSeconds((int)value);
+		    return DateTimeOffset.FromUnixTimeSeconds((long)value);
 	    }
     }
 }
